Add create and delete to tag list
